requirements.txt
telethon>=1.34.0
python-socks[asyncio]>=2.0.0
python-telegram-bot>=21.0
